refactor: Dio Web Api Client
